pyproject.toml
[tool.poetry]
name = "aban-tether"
version = "0.1.0"
description = "AbanTether assignment challenge."
authors = ["Yaser Khahani <yaserkh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = ">=5.0.3,<5.1"
djangorestframework = ">=3.14.0,<3.15"
psycopg2 = ">=2.9.9,<2.10"
pydantic = ">=2.6.4,<2.7"
django-redis = ">=5.4.0,<5.5"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.3.2,<0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
